feat(dashboard): design system tokens, embedded resource split, pre-compressed root
Three intertwined changes that all touch DashboardHtml.cs together:

- Inline the FlowOrchestrator design system token palette (--fo-terracotta,
  --fo-parchment, --fo-ivory, --fo-warm-sand, etc., plus type/radius/ring
  tokens) sourced from .claude/designs/FlowOrchestrator Design System/
  colors_and_type.css. Existing --bg / --accent / --surface tokens kept
  as back-compat aliases so the 197 var(--*) usages elsewhere in the file
  keep working unchanged. Also pulls Source Serif 4 into the inline Google
  Fonts import for future serif-driven UI.
- Split the dashboard SPA into three embedded resources (Assets/index.html,
  Assets/dashboard.css, Assets/dashboard.js) loaded once at static-init and
  inlined into the cached template. DashboardHtml.cs drops from 2103 to
  131 lines; the rendered HTML is byte-identical to before.
- Pre-compress the rendered root page once at startup into Brotli + Gzip
  byte buffers via the new PrecompressedDashboardPage type. The MapGet
  handler picks the best variant per inbound Accept-Encoding (RFC 7231
  q=0 honored), emits Vary: Accept-Encoding, and avoids per-request CPU.
  Typical 80 KB payload now ~12 KB on the wire.

DESIGN.md gains a §10 Token Reference describing the three-surface sync
convention. New integration tests guard the assembly correctness
(DashboardRootPageTests — 8 facts + 1 theory \xc3\x97 7 cases) and the compression
behaviour (DashboardCompressionTests — 9 facts: encoding selection, q=0
fallback, Vary header, decompression equivalence, ratio assertion).

Tests: 318 dashboard integration tests pass (106 \xc3\x97 net8/9/10).
Build: 0 errors / 0 warnings.
